Update Address decoding to fix Litecoin decode bug
Litecoin segwit addresses start with ltc1, however there is about a 1 in 40,000
chance that a legacy L address will also start with ltc1. When this happens the
address decoding fails as it make the decision to parse as segwit or parse as
legacy based on the prefix. In this case it tries and fails to parse as segwit
since the prefix is ltc.

This commit makes a small change to the parse logic to check both the prefix
and the encoding when deciding whether to parse as segwit or legacy.

Tests are updated to test for this case.